+++ tables/c_imt.c
@@ -151,7 +151,7 @@ IRAFPointer imt         i: file name template descriptor
         free (imt_descr->pattern);
 
     if (imt_descr->files) {
-        for (size_t i = 0; i < imt_descr->nfiles;  i++) {
+        for (int i = 0; i < imt_descr->nfiles;  i++) {
             if (imt_descr->files[i])
                 free (imt_descr->files[i]);
         }
